requirements.txt
kafka-python==2.1.5
psycopg2-binary==2.9.9
requests==2.32.3
boto3==1.34.0
geopandas==0.14.3
shapely==2.0.3
folium==0.16.0
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.13.0
jupyterlab==4.1.5
